+++ src/Irc.FSharp/AssemblyInfo.fs
@@ -6,8 +6,8 @@ open System.Runtime.InteropServices
 [<assembly: AssemblyTitleAttribute("Irc.FSharp")>]
 [<assembly: AssemblyProductAttribute("Irc.FSharp")>]
 [<assembly: AssemblyDescriptionAttribute("An IRC client library for F#.")>]
-[<assembly: AssemblyVersionAttribute("0.1.1")>]
-[<assembly: AssemblyFileVersionAttribute("0.1.1")>]
+[<assembly: AssemblyVersionAttribute("0.2.0")>]
+[<assembly: AssemblyFileVersionAttribute("0.2.0")>]
 [<assembly: GuidAttribute("694ab3b0-8929-4f78-ab72-55f29eb48a36")>]
 do ()
 
@@ -15,6 +15,6 @@ module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "Irc.FSharp"
     let [<Literal>] AssemblyProduct = "Irc.FSharp"
     let [<Literal>] AssemblyDescription = "An IRC client library for F#."
-    let [<Literal>] AssemblyVersion = "0.1.1"
-    let [<Literal>] AssemblyFileVersion = "0.1.1"
+    let [<Literal>] AssemblyVersion = "0.2.0"
+    let [<Literal>] AssemblyFileVersion = "0.2.0"
     let [<Literal>] Guid = "694ab3b0-8929-4f78-ab72-55f29eb48a36"